% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_agents}
\alias{parse_agents}
\title{Parse User Agents}
\usage{
parse_agents(user_agents, fields = NULL)
}
\arguments{
\item{user_agents}{a vector of user agents}

\item{fields}{a vector of the fields you want to return - options are "device", "os", "os_major", "os_minor",
"os_patch","os_patch_minor", "browser", "browser_major", "browser_minor", "browser_patch" and "browser_patch_minor",
where minor/major refer to version numbers. By default, all fields are returned.}
}
\value{
a data.frame consisting of those fields you wanted to return.
}
\description{
Take a vector of user agents and parse
them into their component parts.
}
\seealso{
\code{\link{update_regexes}}, for updating the regular expressions file.
}

