% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_r_agents}
\alias{parse_r_agents}
\title{Parse R User Agents}
\usage{
parse_r_agents(user_agents)
}
\arguments{
\item{user_agents}{a vector of user agents}
}
\description{
parses the user agents used by the R programming language
in making web requests, returning the R version, architecture and platform.
}
\examples{
parse_r_agents("R (3.2.0 x86_64-pc-linux-gnu x86_64 linux-gnu)")
}
\seealso{
\code{\link{parse_agents}} for generalised user agent parsing, and
code{\link[urltools]{url_decode}} for removing percentage-encoding from user agents
before processing them.
}

