% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{uaparser}
\alias{uaparser}
\alias{uaparser-package}
\title{uaparser}
\usage{
uaparser(user_agents, fields = NULL)
}
\arguments{
\item{user_agents}{a vector of user agents}

\item{fields}{a vector of the fields you want to return - options are "device", "os", "os_major", "os_minor",
"os_patch","os_patch_minor", "browser", "browser_major", "browser_minor", "browser_patch" and "browser_patch_minor",
where minor/major refer to version numbers.}
}
\value{
a data.frame consisting of those fields you wanted to return. In the event that no "fields"
parameter was set by the user, all fields will be returned.
}
\description{
User agent parser
}
\details{
uaparser (both the function and the package) are the R port of
\href{https://github.com/ua-parser/uap-core}{tobie's ua-parser}. Rather than implementing
the parser entirely in R (which would be slow for the kind of operations that it performed),
this is largely a wrapper around the C++ implementation.
}
\author{
Oliver Keyes <okeyes@wikimedia.org>
}

