% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse.R
\name{update_regexes}
\alias{update_regexes}
\title{Update Regular Expressions}
\usage{
update_regexes()
}
\description{
\code{uaparser} includes a set of regular expressions used to parse user agents.
These will be updated on every new version of the package, but can still get out of date
in the meantime. \code{update_regexes} replaces the version of the regex file with
the latest version from the central UA-Parser repository.
}
\seealso{
\code{\link{parse_agents}}, which relies on the underlying regex file.
}

